moved logging timestamp converter to its own tool project
