+++ kieker-tools/convert-logging-timestamp/build.gradle
@@ -1,0 +1,27 @@
+apply plugin: 'application'
+
+mainClassName='kieker.tools.logging.timestamp.converter.LoggingTimestampConverterTool'
+
+dependencies {
+	clover 'org.openclover:clover:4.2.0'
+
+    compile project(":kieker-common")
+	compile project(":kieker-tools")
+	compile "com.beust:jcommander:1.60"
+	
+	testCompile project (path: ':kieker-common', configuration: 'testArchives')
+	testCompile project (path: ':kieker-tools', configuration: 'testArchives')
+}
+
+clover {
+	report {
+		html=true
+	}
+	
+}
+
+eclipse {
+	classpath {
+		defaultOutputDir = file('build-eclipse')
+	}
+}
