+++ settings.gradle
@@ -8,6 +8,7 @@ include 'kieker-tools'
 include 'kieker-tools:collector'
 include 'kieker-tools:log-replayer'
 include 'kieker-tools:opad'
+include 'kieker-tools:convert-logging-timestamp'
 include 'kieker-tools:resource-monitor'
 include 'kieker-tools:trace-analysis'
 include 'kieker-tools:trace-analysis-gui'
